+++ tests/testthat/test_XcmsExperimentHdf5-functions.R
@@ -247,6 +247,9 @@ test_that(".h5_chrom_peak_ms_levels works", {
 test_that(".h5_subset_xcms_experiment works", {
     expect_error(.h5_subset_xcms_experiment(
         xmseg_full_h5 , c(2, 4, -1, 3, -4)), "Mixing positive")
+    expect_error(.h5_subset_xcms_experiment(
+        xmseg_full_h5, c(1, 2), keepAdjustedRtime = FALSE,
+        keepChromPeaks = TRUE), "Reverting retention times")
 
     a <- new("XcmsExperimentHdf5")
     res <- .h5_subset_xcms_experiment(a)
@@ -271,8 +274,8 @@ test_that(".h5_subset_xcms_experiment works", {
     expect_false(hasChromPeaks(res))
     expect_equal(res@processHistory, a@processHistory)
 
-    res <- .h5_subset_xcms_experiment(xmseg_full_h5, c(3, 1),
-                                      keepFeatures = TRUE)
+    res <- .h5_subset_xcms_experiment(
+        xmseg_full_h5, c(3, 1), keepFeatures = TRUE, keepAdjustedRtime = TRUE)
     expect_true(hasFeatures(res))
     expect_true(length(res) == 2)
     expect_equal(featureDefinitions(res), featureDefinitions(xmseg_full_h5))
@@ -281,14 +284,15 @@ test_that(".h5_subset_xcms_experiment works", {
     expect_equal(a[a[, "sample"] == 1, 1:10], b[b[, "sample"] == 3, 1:10])
     expect_equal(a[a[, "sample"] == 2, 1:10], b[b[, "sample"] == 1, 1:10])
 
-    res <- .h5_subset_xcms_experiment(xmseg_full_h5, c(3, 1),
-                                      keepFeatures = FALSE)
+    res <- .h5_subset_xcms_experiment(
+        xmseg_full_h5, c(3, 1), keepFeatures = FALSE, keepAdjustedRtime = TRUE)
     expect_false(hasFeatures(res))
     expect_true(length(res) == 2)
 
     res <- .h5_subset_xcms_experiment(xmseg_full_h5, c(3, 1),
                                       keepFeatures = TRUE,
-                                      keepChromPeaks = FALSE)
+                                      keepChromPeaks = FALSE,
+                                      keepAdjustedRtime = TRUE)
     expect_false(hasFeatures(res))
     expect_false(hasChromPeaks(res))
     expect_true(length(res) == 2)